Reduce update queries if pk is uuid
When the column of bitemporal_id_key is uuid, the uuid can be generated and specified before insertion into the DB.
In this case, there is no need to update bitemporal_id_key in after_create, so the update query can be reduced by one query.

Note that for types other than uuid, it is not possible to specify the next pk before inserting it into the DB, so the current logic is maintained.